Ago-Dic-2021/hernandez-saucedo-damian-rafael/Extraordinario/app/requirements.txt
mysql-connector-python==8.0.27
requests==2.25.1
pytest==6.2.5